fix(max-test): resolve issue #26 - Fix bugs in calculator.py and tests.py
