+++ test_calculator.py
@@ -1,0 +1,26 @@
+import pytest
+from calculator import calculate_average, process_numbers, divide
+
+def test_divide_by_zero_raises_value_error():
+    with pytest.raises(ValueError):
+        divide(10, 0)
+
+def test_calculate_average_empty_list_raises_value_error():
+    with pytest.raises(ValueError):
+        calculate_average([])
+
+def test_process_numbers_empty_list_returns_empty_dict():
+    assert process_numbers([]) == {}
+
+def test_divide():
+    assert divide(10, 2) == 5.0
+
+def test_average():
+    result = calculate_average([1, 2, 3, 4, 5])
+    assert result == 3.0
+
+def test_process_numbers():
+    result = process_numbers([1, 2, 3])
+    assert result["average"] == 2.0
+    assert result["sum"] == 6
+    assert result["count"] == 3
